#217: Consolidate EntityType
