+++ src/components/Dashboard/Profile/ProfileLayout.tsx
@@ -1,12 +1,11 @@
 "use client";
 import { DashboardLayout } from "@/components/Layout";
 import { ProfileController } from "./ProfileController";
-
-export type ProfileEntityType = "volunteer" | "opportunity" | "agent";
+import { EntityType } from "./types/types";
 
 interface ProfileLayoutProps {
   entityId: string;
-  entityType: ProfileEntityType;
+  entityType: EntityType;
 }
 
 export default function ProfileLayout({ entityId, entityType }: ProfileLayoutProps) {
