+++ src/components/Dashboard/Profile/sections/ContactDetails/ContactDetails.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { ProfileEntityProps } from "@/components/Dashboard/Profile/types/types";
+import { ProfileEntityProps } from "@/components/Dashboard/Profile/types";
 import { forwardRef } from "react";
 import { EditableSectionRef } from "../shared/types";
 import { OpportunityContactDetails } from "./opportunity";
