+++ src/components/Dashboard/Profile/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./agent";
+export * from "./types";
