+++ src/hooks/useCreateComment.ts
@@ -1,3 +1,4 @@
+import { EntityType } from "@/components/Dashboard/Profile/types";
 import { apiPathComment } from "@/config/constants";
 import { useMutationQuery } from "@/hooks";
 import { EntityTableName, Id } from "need4deed-sdk";
@@ -7,8 +8,6 @@ type CreateCommentData = {
   entityType: EntityTableName | string;
   entityId: Id;
 };
-
-type EntityType = "volunteer" | "opportunity" | "agent";
 
 export const useCreateComment = (entityId: Id, entityType: EntityType = "volunteer") => {
   return useMutationQuery<CreateCommentData, { message: string }>({
