+++ src/hooks/useDeleteComment.ts
@@ -1,8 +1,7 @@
+import { EntityType } from "@/components/Dashboard/Profile/types";
 import { apiPathComment } from "@/config/constants";
 import { useMutationQuery } from "@/hooks";
 import { Id } from "need4deed-sdk";
-
-type EntityType = "volunteer" | "opportunity" | "agent";
 
 export const useDeleteComment = (entityId: Id, commentId: number, entityType: EntityType = "volunteer") => {
   return useMutationQuery<unknown, { message: string }>({
